AIoT_sale/requirements.txt
allure-pytest==2.9.43
allure-python-commons==2.9.43
attrs==21.2.0
cachetools==5.0.0
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
colorlog==6.6.0
cssselect==1.1.0
cssutils==2.4.0
execnet==1.9.0
idna==2.10
iniconfig==1.1.1
lxml==4.8.0
packaging==21.0
pluggy==1.0.0
premailer==3.10.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
pytest-forked==1.4.0
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-rerunfailures==10.2
pytest-xdist==2.5.0
PyYAML==6.0
requests==2.25.1
selenium==3.141.0
six==1.16.0
toml==0.10.2
urllib3==1.26.6
yagmail==0.15.277